Move Writer into Control.Algebra.Internal.
